refactor: userlist 2 rpc && fix: register have not status && fix: when total eq 0 have not data

+++ app/user/api/internal/types/types.go
@@ -40,8 +40,8 @@ type UserListReq struct {
 	Phone    string `json:"phone,optional"`
 	Nickname string `json:"nickname,optional"`
 	Status   int64  `json:"status,optional"`
-	PageNo   int    `json:"pageNo,optional"`
-	PageSize int    `json:"pageSize,optional"`
+	PageNo   int64  `json:"pageNo,optional"`
+	PageSize int64  `json:"pageSize,optional"`
 }
 
 type UserListRsp struct {
